Rename prompt_execution to execute
